Refactor config handling to registry-common
These classes are needed elsewhere and cannot be called from a sprint boot app

refs nasa-pds/validate#895
