+++ pom.xml
@@ -113,7 +113,7 @@
     <dependency>
       <groupId>gov.nasa.pds</groupId>
       <artifactId>registry-common</artifactId>
-      <version>2.0.0</version>
+      <version>2.1.0-SNAPSHOT</version>
     </dependency>
     <!-- rich XML parsing -->
     <dependency>
