--- src/main/java/gov/nasa/pds/harvest/cfg/AutogenFieldsType.java
@@ -1,103 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v4.0.3 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-//
-
-
-package gov.nasa.pds.harvest.cfg;
-
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlAttribute;
-import jakarta.xml.bind.annotation.XmlType;
-
-
-/**
- * 
- *         Determines how items found in the products should be indexed in the
- *         registry. Do not define this tag for normal, expected behavior.
- *       
- * 
- * <p>Java class for autogen_fields_type complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>{@code
- * <complexType name="autogen_fields_type">
- *   <complexContent>
- *     <restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       <sequence>
- *         <element name="classFilter" type="{}filter_type" minOccurs="0"/>
- *       </sequence>
- *       <attribute name="generate" type="{http://www.w3.org/2001/XMLSchema}boolean" default="true" />
- *     </restriction>
- *   </complexContent>
- * </complexType>
- * }</pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "autogen_fields_type", propOrder = {
-    "classFilter"
-})
-public class AutogenFieldsType {
-
-    protected FilterType classFilter;
-    @XmlAttribute(name = "generate")
-    protected Boolean generate;
-
-    /**
-     * Gets the value of the classFilter property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link FilterType }
-     *     
-     */
-    public FilterType getClassFilter() {
-        return classFilter;
-    }
-
-    /**
-     * Sets the value of the classFilter property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link FilterType }
-     *     
-     */
-    public void setClassFilter(FilterType value) {
-        this.classFilter = value;
-    }
-
-    /**
-     * Gets the value of the generate property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Boolean }
-     *     
-     */
-    public boolean isGenerate() {
-        if (generate == null) {
-            return true;
-        } else {
-            return generate;
-        }
-    }
-
-    /**
-     * Sets the value of the generate property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Boolean }
-     *     
-     */
-    public void setGenerate(Boolean value) {
-        this.generate = value;
-    }
-
-}
