--- src/main/java/gov/nasa/pds/harvest/cfg/BundleType.java
@@ -1,188 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v4.0.3 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-//
-
-
-package gov.nasa.pds.harvest.cfg;
-
-import java.util.ArrayList;
-import java.util.List;
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlAttribute;
-import jakarta.xml.bind.annotation.XmlSchemaType;
-import jakarta.xml.bind.annotation.XmlType;
-import jakarta.xml.bind.annotation.adapters.NormalizedStringAdapter;
-import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-
-
-/**
- * 
- *         Describe the bundle to be processed.
- * 
- *         @dir: must supply the directory that contains the bundle XML file
- *         @versions: a comma separated list of values but "all", the default,
- *                    will ingest all versions.
- * 
- *         collection: list of collections to include but if not specified, then
- *                     all collections found and referenced by the bundle will be
- *                     included.
- *         product: list of product directories. Reason???
- *       
- * 
- * <p>Java class for bundle_type complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>{@code
- * <complexType name="bundle_type">
- *   <complexContent>
- *     <restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       <sequence>
- *         <element name="collection" type="{}collection_type" maxOccurs="unbounded" minOccurs="0"/>
- *         <element name="product" type="{}product_type" maxOccurs="unbounded" minOccurs="0"/>
- *       </sequence>
- *       <attribute name="dir" use="required" type="{http://www.w3.org/2001/XMLSchema}normalizedString" />
- *       <attribute name="versions" type="{http://www.w3.org/2001/XMLSchema}normalizedString" default="all" />
- *     </restriction>
- *   </complexContent>
- * </complexType>
- * }</pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "bundle_type", propOrder = {
-    "collection",
-    "product"
-})
-public class BundleType {
-
-    protected List<CollectionType> collection;
-    protected List<ProductType> product;
-    @XmlAttribute(name = "dir", required = true)
-    @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
-    @XmlSchemaType(name = "normalizedString")
-    protected String dir;
-    @XmlAttribute(name = "versions")
-    @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
-    @XmlSchemaType(name = "normalizedString")
-    protected String versions;
-
-    /**
-     * Gets the value of the collection property.
-     * 
-     * <p>
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the Jakarta XML Binding object.
-     * This is why there is not a {@code set} method for the collection property.
-     * 
-     * <p>
-     * For example, to add a new item, do as follows:
-     * <pre>
-     *    getCollection().add(newItem);
-     * </pre>
-     * 
-     * 
-     * <p>
-     * Objects of the following type(s) are allowed in the list
-     * {@link CollectionType }
-     * 
-     * 
-     * @return
-     *     The value of the collection property.
-     */
-    public List<CollectionType> getCollection() {
-        if (collection == null) {
-            collection = new ArrayList<>();
-        }
-        return this.collection;
-    }
-
-    /**
-     * Gets the value of the product property.
-     * 
-     * <p>
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the Jakarta XML Binding object.
-     * This is why there is not a {@code set} method for the product property.
-     * 
-     * <p>
-     * For example, to add a new item, do as follows:
-     * <pre>
-     *    getProduct().add(newItem);
-     * </pre>
-     * 
-     * 
-     * <p>
-     * Objects of the following type(s) are allowed in the list
-     * {@link ProductType }
-     * 
-     * 
-     * @return
-     *     The value of the product property.
-     */
-    public List<ProductType> getProduct() {
-        if (product == null) {
-            product = new ArrayList<>();
-        }
-        return this.product;
-    }
-
-    /**
-     * Gets the value of the dir property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getDir() {
-        return dir;
-    }
-
-    /**
-     * Sets the value of the dir property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setDir(String value) {
-        this.dir = value;
-    }
-
-    /**
-     * Gets the value of the versions property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getVersions() {
-        if (versions == null) {
-            return "all";
-        } else {
-            return versions;
-        }
-    }
-
-    /**
-     * Sets the value of the versions property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setVersions(String value) {
-        this.versions = value;
-    }
-
-}
