--- src/main/java/gov/nasa/pds/harvest/cfg/CollectionType.java
@@ -1,119 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v4.0.3 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-//
-
-
-package gov.nasa.pds.harvest.cfg;
-
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlAttribute;
-import jakarta.xml.bind.annotation.XmlSchemaType;
-import jakarta.xml.bind.annotation.XmlType;
-import jakarta.xml.bind.annotation.adapters.NormalizedStringAdapter;
-import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-
-
-/**
- * 
- *         The collection's LID or LIDVID.
- * 
- *         DEV notes:
- * 
- *         Unfortunately, cannot define mutually exclusive attributes
- *         without schematron and trying to avoid schematron.
- * 
- *         It may be better to ditch either lid or lidvid then if :: is found
- *         treat it as lidvid otherwise lid. Doing that would allow for the
- *         attribute lid as required but accept lidvids.
- *       
- * 
- * <p>Java class for collection_type complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>{@code
- * <complexType name="collection_type">
- *   <complexContent>
- *     <restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       <attribute name="lid" type="{http://www.w3.org/2001/XMLSchema}normalizedString" default="" />
- *       <attribute name="lidvid" type="{http://www.w3.org/2001/XMLSchema}normalizedString" default="" />
- *     </restriction>
- *   </complexContent>
- * </complexType>
- * }</pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "collection_type")
-public class CollectionType {
-
-    @XmlAttribute(name = "lid")
-    @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
-    @XmlSchemaType(name = "normalizedString")
-    protected String lid;
-    @XmlAttribute(name = "lidvid")
-    @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
-    @XmlSchemaType(name = "normalizedString")
-    protected String lidvid;
-
-    /**
-     * Gets the value of the lid property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getLid() {
-        if (lid == null) {
-            return "";
-        } else {
-            return lid;
-        }
-    }
-
-    /**
-     * Sets the value of the lid property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setLid(String value) {
-        this.lid = value;
-    }
-
-    /**
-     * Gets the value of the lidvid property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getLidvid() {
-        if (lidvid == null) {
-            return "";
-        } else {
-            return lidvid;
-        }
-    }
-
-    /**
-     * Sets the value of the lidvid property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setLidvid(String value) {
-        this.lidvid = value;
-    }
-
-}
