--- src/main/java/gov/nasa/pds/harvest/cfg/FileInfoType.java
@@ -1,181 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v4.0.3 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-//
-
-
-package gov.nasa.pds.harvest.cfg;
-
-import java.util.ArrayList;
-import java.util.List;
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlAttribute;
-import jakarta.xml.bind.annotation.XmlElement;
-import jakarta.xml.bind.annotation.XmlType;
-
-
-/**
- * 
- *         Used to rename file locations from a local space to the install space.
- *         For instance, change /home/username/pdsfiles to
- *         https://public.space.com/pdsfiles,
- * 
- *         @processDataFiles:
- *         @storeJsonLabels:
- *         @storeLabels:
- *       
- * 
- * <p>Java class for file_info_type complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>{@code
- * <complexType name="file_info_type">
- *   <complexContent>
- *     <restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       <sequence>
- *         <element name="fileRef" type="{}file_ref_type" maxOccurs="unbounded"/>
- *       </sequence>
- *       <attribute name="processDataFiles" type="{http://www.w3.org/2001/XMLSchema}boolean" default="true" />
- *       <attribute name="storeJsonLabels" type="{http://www.w3.org/2001/XMLSchema}boolean" default="true" />
- *       <attribute name="storeLabels" type="{http://www.w3.org/2001/XMLSchema}boolean" default="true" />
- *     </restriction>
- *   </complexContent>
- * </complexType>
- * }</pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "file_info_type", propOrder = {
-    "fileRef"
-})
-public class FileInfoType {
-
-    @XmlElement(required = true)
-    protected List<FileRefType> fileRef;
-    @XmlAttribute(name = "processDataFiles")
-    protected Boolean processDataFiles;
-    @XmlAttribute(name = "storeJsonLabels")
-    protected Boolean storeJsonLabels;
-    @XmlAttribute(name = "storeLabels")
-    protected Boolean storeLabels;
-
-    /**
-     * Gets the value of the fileRef property.
-     * 
-     * <p>
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the Jakarta XML Binding object.
-     * This is why there is not a {@code set} method for the fileRef property.
-     * 
-     * <p>
-     * For example, to add a new item, do as follows:
-     * <pre>
-     *    getFileRef().add(newItem);
-     * </pre>
-     * 
-     * 
-     * <p>
-     * Objects of the following type(s) are allowed in the list
-     * {@link FileRefType }
-     * 
-     * 
-     * @return
-     *     The value of the fileRef property.
-     */
-    public List<FileRefType> getFileRef() {
-        if (fileRef == null) {
-            fileRef = new ArrayList<>();
-        }
-        return this.fileRef;
-    }
-
-    /**
-     * Gets the value of the processDataFiles property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Boolean }
-     *     
-     */
-    public boolean isProcessDataFiles() {
-        if (processDataFiles == null) {
-            return true;
-        } else {
-            return processDataFiles;
-        }
-    }
-
-    /**
-     * Sets the value of the processDataFiles property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Boolean }
-     *     
-     */
-    public void setProcessDataFiles(Boolean value) {
-        this.processDataFiles = value;
-    }
-
-    /**
-     * Gets the value of the storeJsonLabels property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Boolean }
-     *     
-     */
-    public boolean isStoreJsonLabels() {
-        if (storeJsonLabels == null) {
-            return true;
-        } else {
-            return storeJsonLabels;
-        }
-    }
-
-    /**
-     * Sets the value of the storeJsonLabels property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Boolean }
-     *     
-     */
-    public void setStoreJsonLabels(Boolean value) {
-        this.storeJsonLabels = value;
-    }
-
-    /**
-     * Gets the value of the storeLabels property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Boolean }
-     *     
-     */
-    public boolean isStoreLabels() {
-        if (storeLabels == null) {
-            return true;
-        } else {
-            return storeLabels;
-        }
-    }
-
-    /**
-     * Sets the value of the storeLabels property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Boolean }
-     *     
-     */
-    public void setStoreLabels(Boolean value) {
-        this.storeLabels = value;
-    }
-
-}
