--- src/main/java/gov/nasa/pds/harvest/cfg/Harvest.java
@@ -1,48 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v4.0.3 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-//
-
-
-package gov.nasa.pds.harvest.cfg;
-
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlRootElement;
-import jakarta.xml.bind.annotation.XmlType;
-
-
-/**
- * 
- *           This terrible construct is so that xjc can autodetect this as the
- *           root node for processing. Many things would be better but this is
- *           the most workable solution especially if the making of the binding
- *           code is automated in the pom. The only other real solution is to
- *           modify one of the classes generated by hand.
- *         
- * 
- * <p>Java class for anonymous complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>{@code
- * <complexType>
- *   <complexContent>
- *     <extension base="{}harvest_configuration_type">
- *     </extension>
- *   </complexContent>
- * </complexType>
- * }</pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "")
-@XmlRootElement(name = "harvest")
-public class Harvest
-    extends HarvestConfigurationType
-{
-
-
-}
