--- src/main/java/gov/nasa/pds/harvest/cfg/LoadType.java
@@ -1,126 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v4.0.3 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-//
-
-
-package gov.nasa.pds.harvest.cfg;
-
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlType;
-
-
-/**
- * 
- *         DEV notes:
- * 
- *         Had to bury this one deeper because cannot have a choice in an all.
- *       
- * 
- * <p>Java class for load_type complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>{@code
- * <complexType name="load_type">
- *   <complexContent>
- *     <restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       <choice>
- *         <element name="bundles" type="{}bundles_type"/>
- *         <element name="directories" type="{}dirs_type"/>
- *         <element name="files" type="{}files_type"/>
- *       </choice>
- *     </restriction>
- *   </complexContent>
- * </complexType>
- * }</pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "load_type", propOrder = {
-    "bundles",
-    "directories",
-    "files"
-})
-public class LoadType {
-
-    protected BundlesType bundles;
-    protected DirsType directories;
-    protected FilesType files;
-
-    /**
-     * Gets the value of the bundles property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link BundlesType }
-     *     
-     */
-    public BundlesType getBundles() {
-        return bundles;
-    }
-
-    /**
-     * Sets the value of the bundles property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link BundlesType }
-     *     
-     */
-    public void setBundles(BundlesType value) {
-        this.bundles = value;
-    }
-
-    /**
-     * Gets the value of the directories property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link DirsType }
-     *     
-     */
-    public DirsType getDirectories() {
-        return directories;
-    }
-
-    /**
-     * Sets the value of the directories property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link DirsType }
-     *     
-     */
-    public void setDirectories(DirsType value) {
-        this.directories = value;
-    }
-
-    /**
-     * Gets the value of the files property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link FilesType }
-     *     
-     */
-    public FilesType getFiles() {
-        return files;
-    }
-
-    /**
-     * Sets the value of the files property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link FilesType }
-     *     
-     */
-    public void setFiles(FilesType value) {
-        this.files = value;
-    }
-
-}
