--- src/main/java/gov/nasa/pds/harvest/cfg/ObjectFactory.java
@@ -1,208 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v4.0.3 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-//
-
-
-package gov.nasa.pds.harvest.cfg;
-
-import jakarta.xml.bind.annotation.XmlRegistry;
-
-
-/**
- * This object contains factory methods for each 
- * Java content interface and Java element interface 
- * generated in the gov.nasa.pds.harvest.cfg package. 
- * <p>An ObjectFactory allows you to programmatically 
- * construct new instances of the Java representation 
- * for XML content. The Java representation of XML 
- * content can consist of schema derived interfaces 
- * and classes representing the binding of schema 
- * type definitions, element declarations and model 
- * groups.  Factory methods for each of these are 
- * provided in this class.
- * 
- */
-@XmlRegistry
-public class ObjectFactory {
-
-
-    /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: gov.nasa.pds.harvest.cfg
-     * 
-     */
-    public ObjectFactory() {
-    }
-
-    /**
-     * Create an instance of {@link Harvest }
-     * 
-     * @return
-     *     the new instance of {@link Harvest }
-     */
-    public Harvest createHarvest() {
-        return new Harvest();
-    }
-
-    /**
-     * Create an instance of {@link HarvestConfigurationType }
-     * 
-     * @return
-     *     the new instance of {@link HarvestConfigurationType }
-     */
-    public HarvestConfigurationType createHarvestConfigurationType() {
-        return new HarvestConfigurationType();
-    }
-
-    /**
-     * Create an instance of {@link AutogenFieldsType }
-     * 
-     * @return
-     *     the new instance of {@link AutogenFieldsType }
-     */
-    public AutogenFieldsType createAutogenFieldsType() {
-        return new AutogenFieldsType();
-    }
-
-    /**
-     * Create an instance of {@link LoadType }
-     * 
-     * @return
-     *     the new instance of {@link LoadType }
-     */
-    public LoadType createLoadType() {
-        return new LoadType();
-    }
-
-    /**
-     * Create an instance of {@link FileInfoType }
-     * 
-     * @return
-     *     the new instance of {@link FileInfoType }
-     */
-    public FileInfoType createFileInfoType() {
-        return new FileInfoType();
-    }
-
-    /**
-     * Create an instance of {@link FilterType }
-     * 
-     * @return
-     *     the new instance of {@link FilterType }
-     */
-    public FilterType createFilterType() {
-        return new FilterType();
-    }
-
-    /**
-     * Create an instance of {@link ReferencesType }
-     * 
-     * @return
-     *     the new instance of {@link ReferencesType }
-     */
-    public ReferencesType createReferencesType() {
-        return new ReferencesType();
-    }
-
-    /**
-     * Create an instance of {@link RegistryType }
-     * 
-     * @return
-     *     the new instance of {@link RegistryType }
-     */
-    public RegistryType createRegistryType() {
-        return new RegistryType();
-    }
-
-    /**
-     * Create an instance of {@link XpathMapsType }
-     * 
-     * @return
-     *     the new instance of {@link XpathMapsType }
-     */
-    public XpathMapsType createXpathMapsType() {
-        return new XpathMapsType();
-    }
-
-    /**
-     * Create an instance of {@link BundleType }
-     * 
-     * @return
-     *     the new instance of {@link BundleType }
-     */
-    public BundleType createBundleType() {
-        return new BundleType();
-    }
-
-    /**
-     * Create an instance of {@link BundlesType }
-     * 
-     * @return
-     *     the new instance of {@link BundlesType }
-     */
-    public BundlesType createBundlesType() {
-        return new BundlesType();
-    }
-
-    /**
-     * Create an instance of {@link CollectionType }
-     * 
-     * @return
-     *     the new instance of {@link CollectionType }
-     */
-    public CollectionType createCollectionType() {
-        return new CollectionType();
-    }
-
-    /**
-     * Create an instance of {@link DirsType }
-     * 
-     * @return
-     *     the new instance of {@link DirsType }
-     */
-    public DirsType createDirsType() {
-        return new DirsType();
-    }
-
-    /**
-     * Create an instance of {@link FileRefType }
-     * 
-     * @return
-     *     the new instance of {@link FileRefType }
-     */
-    public FileRefType createFileRefType() {
-        return new FileRefType();
-    }
-
-    /**
-     * Create an instance of {@link FilesType }
-     * 
-     * @return
-     *     the new instance of {@link FilesType }
-     */
-    public FilesType createFilesType() {
-        return new FilesType();
-    }
-
-    /**
-     * Create an instance of {@link ProductType }
-     * 
-     * @return
-     *     the new instance of {@link ProductType }
-     */
-    public ProductType createProductType() {
-        return new ProductType();
-    }
-
-    /**
-     * Create an instance of {@link XpathMapType }
-     * 
-     * @return
-     *     the new instance of {@link XpathMapType }
-     */
-    public XpathMapType createXpathMapType() {
-        return new XpathMapType();
-    }
-
-}
