--- src/main/java/gov/nasa/pds/harvest/cfg/ReferencesType.java
@@ -1,72 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v4.0.3 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-//
-
-
-package gov.nasa.pds.harvest.cfg;
-
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlAttribute;
-import jakarta.xml.bind.annotation.XmlType;
-
-
-/**
- * 
- *         Seems like a forgotten idea
- *       
- * 
- * <p>Java class for references_type complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>{@code
- * <complexType name="references_type">
- *   <complexContent>
- *     <restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       <attribute name="primaryOnly" type="{http://www.w3.org/2001/XMLSchema}boolean" default="false" />
- *     </restriction>
- *   </complexContent>
- * </complexType>
- * }</pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "references_type")
-public class ReferencesType {
-
-    @XmlAttribute(name = "primaryOnly")
-    protected Boolean primaryOnly;
-
-    /**
-     * Gets the value of the primaryOnly property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Boolean }
-     *     
-     */
-    public boolean isPrimaryOnly() {
-        if (primaryOnly == null) {
-            return false;
-        } else {
-            return primaryOnly;
-        }
-    }
-
-    /**
-     * Sets the value of the primaryOnly property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Boolean }
-     *     
-     */
-    public void setPrimaryOnly(Boolean value) {
-        this.primaryOnly = value;
-    }
-
-}
