--- src/main/java/gov/nasa/pds/harvest/cfg/XpathMapType.java
@@ -1,113 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v4.0.3 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-//
-
-
-package gov.nasa.pds.harvest.cfg;
-
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlAttribute;
-import jakarta.xml.bind.annotation.XmlSchemaType;
-import jakarta.xml.bind.annotation.XmlType;
-import jakarta.xml.bind.annotation.adapters.NormalizedStringAdapter;
-import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-
-
-/**
- * 
- *         Point at an xpath file to apply against PDS4 labels. If @rootElement
- *         is the default value, then the xpath file will be applied to labels.
- *         Otherwise only those labels with matching @rootElement will have the
- *         xpath file applied.
- * 
- *         @filePath: location of the xpath file
- *         @rootElement: name of the root element to apply xpath file with default
- *                       being all.
- *       
- * 
- * <p>Java class for xpath_map_type complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>{@code
- * <complexType name="xpath_map_type">
- *   <complexContent>
- *     <restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       <attribute name="filePath" use="required" type="{http://www.w3.org/2001/XMLSchema}normalizedString" />
- *       <attribute name="rootElement" type="{http://www.w3.org/2001/XMLSchema}normalizedString" default="" />
- *     </restriction>
- *   </complexContent>
- * </complexType>
- * }</pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "xpath_map_type")
-public class XpathMapType {
-
-    @XmlAttribute(name = "filePath", required = true)
-    @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
-    @XmlSchemaType(name = "normalizedString")
-    protected String filePath;
-    @XmlAttribute(name = "rootElement")
-    @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
-    @XmlSchemaType(name = "normalizedString")
-    protected String rootElement;
-
-    /**
-     * Gets the value of the filePath property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getFilePath() {
-        return filePath;
-    }
-
-    /**
-     * Sets the value of the filePath property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setFilePath(String value) {
-        this.filePath = value;
-    }
-
-    /**
-     * Gets the value of the rootElement property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getRootElement() {
-        if (rootElement == null) {
-            return "";
-        } else {
-            return rootElement;
-        }
-    }
-
-    /**
-     * Sets the value of the rootElement property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setRootElement(String value) {
-        this.rootElement = value;
-    }
-
-}
