--- src/main/java/gov/nasa/pds/harvest/cfg/XpathMapsType.java
@@ -1,114 +1,0 @@
-//
-// This file was generated by the Eclipse Implementation of JAXB, v4.0.3 
-// See https://eclipse-ee4j.github.io/jaxb-ri 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-//
-
-
-package gov.nasa.pds.harvest.cfg;
-
-import java.util.ArrayList;
-import java.util.List;
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlAttribute;
-import jakarta.xml.bind.annotation.XmlElement;
-import jakarta.xml.bind.annotation.XmlSchemaType;
-import jakarta.xml.bind.annotation.XmlType;
-import jakarta.xml.bind.annotation.adapters.NormalizedStringAdapter;
-import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-
-
-/**
- * 
- *         The xpaths to apply to PDS4 labels for determining harvesting.
- *       
- * 
- * <p>Java class for xpath_maps_type complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>{@code
- * <complexType name="xpath_maps_type">
- *   <complexContent>
- *     <restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       <sequence>
- *         <element name="xpathMap" type="{}xpath_map_type" maxOccurs="unbounded"/>
- *       </sequence>
- *       <attribute name="baseDir" use="required" type="{http://www.w3.org/2001/XMLSchema}normalizedString" />
- *     </restriction>
- *   </complexContent>
- * </complexType>
- * }</pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "xpath_maps_type", propOrder = {
-    "xpathMap"
-})
-public class XpathMapsType {
-
-    @XmlElement(required = true)
-    protected List<XpathMapType> xpathMap;
-    @XmlAttribute(name = "baseDir", required = true)
-    @XmlJavaTypeAdapter(NormalizedStringAdapter.class)
-    @XmlSchemaType(name = "normalizedString")
-    protected String baseDir;
-
-    /**
-     * Gets the value of the xpathMap property.
-     * 
-     * <p>
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the Jakarta XML Binding object.
-     * This is why there is not a {@code set} method for the xpathMap property.
-     * 
-     * <p>
-     * For example, to add a new item, do as follows:
-     * <pre>
-     *    getXpathMap().add(newItem);
-     * </pre>
-     * 
-     * 
-     * <p>
-     * Objects of the following type(s) are allowed in the list
-     * {@link XpathMapType }
-     * 
-     * 
-     * @return
-     *     The value of the xpathMap property.
-     */
-    public List<XpathMapType> getXpathMap() {
-        if (xpathMap == null) {
-            xpathMap = new ArrayList<>();
-        }
-        return this.xpathMap;
-    }
-
-    /**
-     * Gets the value of the baseDir property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getBaseDir() {
-        return baseDir;
-    }
-
-    /**
-     * Sets the value of the baseDir property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setBaseDir(String value) {
-        this.baseDir = value;
-    }
-
-}
