+++ src/main/java/gov/nasa/pds/harvest/cmd/HarvestCmd.java
@@ -5,9 +5,6 @@ import java.io.File;
 import org.apache.commons.cli.CommandLine;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import gov.nasa.pds.harvest.cfg.BundleType;
-import gov.nasa.pds.harvest.cfg.ConfigManager;
-import gov.nasa.pds.harvest.cfg.HarvestConfigurationType;
 import gov.nasa.pds.harvest.crawler.BundleProcessor;
 import gov.nasa.pds.harvest.crawler.CollectionProcessor;
 import gov.nasa.pds.harvest.crawler.Counter;
@@ -20,6 +17,9 @@ import gov.nasa.pds.harvest.util.CounterMap;
 import gov.nasa.pds.harvest.util.PackageIdGenerator;
 import gov.nasa.pds.harvest.util.log.LogUtils;
 import gov.nasa.pds.harvest.util.out.WriterManager;
+import gov.nasa.pds.registry.common.cfg.BundleType;
+import gov.nasa.pds.registry.common.cfg.ConfigManager;
+import gov.nasa.pds.registry.common.cfg.HarvestConfigurationType;
 import gov.nasa.pds.registry.common.meta.BasicMetadataExtractor;
 import gov.nasa.pds.registry.common.util.ArchiveStatus;
