+++ src/main/java/gov/nasa/pds/harvest/crawler/CollectionProcessor.java
@@ -13,12 +13,12 @@ import java.util.stream.Stream;
 
 import gov.nasa.pds.registry.common.util.CloseUtils;
 import org.w3c.dom.Document;
-import gov.nasa.pds.harvest.cfg.BundleType;
-import gov.nasa.pds.harvest.cfg.ConfigManager;
-import gov.nasa.pds.harvest.cfg.HarvestConfigurationType;
 import gov.nasa.pds.harvest.dao.RegistryDao;
 import gov.nasa.pds.harvest.dao.RegistryManager;
 import gov.nasa.pds.harvest.util.xml.XmlIs;
+import gov.nasa.pds.registry.common.cfg.BundleType;
+import gov.nasa.pds.registry.common.cfg.ConfigManager;
+import gov.nasa.pds.registry.common.cfg.HarvestConfigurationType;
 import gov.nasa.pds.registry.common.es.service.CollectionInventoryWriter;
 import gov.nasa.pds.registry.common.meta.CollectionMetadataExtractor;
 import gov.nasa.pds.registry.common.meta.Metadata;
