+++ src/main/java/gov/nasa/pds/harvest/crawler/FilesProcessor.java
@@ -14,13 +14,13 @@ import java.util.function.BiPredicate;
 import java.util.stream.Stream;
 
 import org.w3c.dom.Document;
-import gov.nasa.pds.harvest.cfg.HarvestConfigurationType;
-import gov.nasa.pds.harvest.cfg.ConfigManager;
 import gov.nasa.pds.harvest.dao.RegistryDao;
 import gov.nasa.pds.harvest.dao.RegistryManager;
 import gov.nasa.pds.harvest.util.out.SupplementalWriter;
 import gov.nasa.pds.harvest.util.out.WriterManager;
 import gov.nasa.pds.harvest.util.xml.XmlIs;
+import gov.nasa.pds.registry.common.cfg.ConfigManager;
+import gov.nasa.pds.registry.common.cfg.HarvestConfigurationType;
 import gov.nasa.pds.registry.common.es.service.CollectionInventoryWriter;
 import gov.nasa.pds.registry.common.meta.BundleMetadataExtractor;
 import gov.nasa.pds.registry.common.meta.CollectionMetadataExtractor;
